refactor: 공통으로 사용되는 Name, Price, Quantity 를 common.domain 패키지로 이동
