+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -9,7 +9,7 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import kitchenpos.product.domain.Product;
-import kitchenpos.domain.Quantity;
+import kitchenpos.common.domain.Quantity;
 
 @Entity
 public class MenuProduct {
