+++ src/test/java/kitchenpos/menu/fixture/MenuProductFixture.java
@@ -2,7 +2,7 @@ package kitchenpos.menu.fixture;
 
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.product.domain.Product;
-import kitchenpos.domain.Quantity;
+import kitchenpos.common.domain.Quantity;
 
 public class MenuProductFixture {
