+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -30,7 +30,7 @@ import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
 import kitchenpos.product.domain.Product;
-import kitchenpos.domain.Quantity;
+import kitchenpos.common.domain.Quantity;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
